newAddition

+++ calculator.py
@@ -64,10 +64,10 @@ while True:
         result = mod(float(num1), float(num2))
 
     elif operator == "x+":
-        result = add_mult(float(num1), float(num2), float(num3))
+        result = add_mult(int(num1), int(num2), int(num3))
 
     elif operator == "cubes+":
-        result = add_cubes(float(num1), float(num2))
+        result = add_cubes(int(num1), int(num2))
 
     else:
         result = "Please enter an operator followed by two integers."
